app/requirements.txt
altair
asttokens
attrs
backcall
beautifulsoup4
bleach
blinker
cachetools
certifi
charset-normalizer
chembl-structure-pipeline
click
colorama
contourpy
cycler
decorator
defusedxml
docopt
entrypoints
et-xmlfile
executing
fastjsonschema
fonttools
gitdb
gitpython
idna
imbalanced-learn
imblearn
ipython
jedi
jinja2
joblib
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
jupyterlab-pygments
kiwisolver
lightgbm
markdown-it-py
markupsafe
matplotlib
matplotlib-inline
mdurl
mistune
nbclient
nbconvert
nbformat
numpy
openpyxl
packaging
pandas
pandocfilters
parso
pickleshare
pillow
pipreqs
platformdirs
plotly
prompt-toolkit
protobuf
pure-eval
pyaml
pyarrow
pydeck
pygments
pyparsing
python-dateutil
python-decouple
pytz
pyyaml
pyzmq
rdkit
referencing
requests
rich
rpds-py
scikit-learn
scikit-optimize
scipy
seaborn
six
smmap
soupsieve
stack-data
streamlit
streamlit-aggrid
tenacity
threadpoolctl
tinycss2
toml
toolz
tornado
traitlets
typing-extensions
tzdata
urllib3
watchdog
wcwidth
webencodings
yarg
